Update music.yaml to make play/pause button change icon
Play/Pause button now follows the state of the media play. When playing, its shows | |, when paused, it shows >.
Changed the icon name of the boxed icon on the main display to have _box for consistent naming of the play/pause icons
Changed update_main into update_all to trigger update of the buttons.